Restructured plugin to follow bevy convention better
The plugin is now actually a plugin again, and gets added to the app as normal. There is now proper logging to ensure a user is aware that the plugin needs to be added before `DefaultPlugins` to work correctly.

This also reduced the complexity of the crate, which is nice.